[*] Email validation.     R: The EmailField override is incorectly implemented.     FB: Changed the order of class inheritans and refactored the serializers of specialists.
