+++ src/backend/api/v1/general/fields.py
@@ -1,19 +1,11 @@
 from base64 import b64decode
 
 from django.core.files.base import ContentFile
-from rest_framework.serializers import ImageField, PrimaryKeyRelatedField
+from django.core.validators import EmailValidator
+from rest_framework.fields import EmailField
+from rest_framework.serializers import ImageField
 
-from .serializers import SkillSerializer
-
-
-class SkillField(PrimaryKeyRelatedField):
-    """
-    Поле навыков: запись по первичному ключу,
-    чтение в виде вложенной структуры.
-    """
-
-    def to_representation(self, skill):
-        return SkillSerializer(skill).data
+from apps.general.validators import CustomEmailValidator
 
 
 class Base64ImageField(ImageField):
@@ -25,3 +17,14 @@ class Base64ImageField(ImageField):
             extension = format.split("/")[-1]
             data = ContentFile(b64decode(imgstr), name="temp." + extension)
         return super().to_internal_value(data)
+
+
+class CustomEmailField(EmailField):
+    """Кастомное поле для email."""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.validators = [
+            v for v in self.validators if not isinstance(v, EmailValidator)
+        ]
+        self.validators.append(CustomEmailValidator())
